backend/requirements.txt
# Career Intelligence Assistant - Python Dependencies
# Python 3.11+

# ============================================================================
# Web Framework
# ============================================================================
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
websockets>=12.0

# ============================================================================
# LLM & RAG
# ============================================================================
openai>=1.12.0
llama-index>=0.10.13
llama-index-llms-openai>=0.1.6
llama-index-embeddings-huggingface>=0.1.4
llama-index-graph-stores-neo4j>=0.1.3
llama-index-vector-stores-neo4jvector>=0.1.2

# ============================================================================
# Web Scraping (Scrapy-based - replaces Tavily)
# ============================================================================
scrapy>=2.11.0
httpx>=0.26.0  # Async HTTP client for scraping

# ============================================================================
# Database
# ============================================================================
neo4j>=5.17.0

# ============================================================================
# Document Parsing
# ============================================================================
pymupdf>=1.23.22  # PDF parsing
python-docx>=1.1.0  # DOCX parsing

# ============================================================================
# Embeddings
# ============================================================================
sentence-transformers>=2.3.1
transformers>=4.37.2
torch>=2.2.0

# ============================================================================
# Data Validation
# ============================================================================
pydantic>=2.6.1
pydantic-settings>=2.1.0

# ============================================================================
# Security & Guardrails
# ============================================================================
presidio-analyzer>=2.2.354  # PII detection
presidio-anonymizer>=2.2.354
slowapi>=0.1.9  # Rate limiting
python-magic>=0.4.27
puremagic>=1.20
einops>=0.7.0  # File type detection

# ============================================================================
# Async Utilities
# ============================================================================
httpx>=0.26.0
aiofiles>=23.2.1

# ============================================================================
# Configuration
# ============================================================================
python-dotenv>=1.0.1

# ============================================================================
# Logging & Monitoring
# ============================================================================
structlog>=24.1.0
python-json-logger>=2.0.7

# ============================================================================
# Testing
# ============================================================================
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For TestClient
respx>=0.20.2  # Mock HTTP
pytest-mock>=3.12.0

# ============================================================================
# Development
# ============================================================================
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0
ruff>=0.2.1
